Postgres/New folder/requirements.txt
sentence-transformers==2.7.0
google-generativeai==0.5.0
psycopg2-binary==2.9.9
pgvector==0.2.5
numpy==1.26.4